1. Установлен роутер
2. Добавлены пути к страницам
3. Разбросаны по компонентам страницы
