+++ package.json
@@ -15,6 +15,7 @@
     "nanoid": "^3.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
     "react-window": "^1.8.6",
     "styled-components": "^5.3.3",
