+++ src/components/index.js
@@ -1,1 +1,3 @@
-export * from './message/message'
+export * from './chat/chat'
+export * from './chatList/chatList'
+export * from './profile/profile'
