+++ src/components/profile/profile.js
@@ -1,0 +1,7 @@
+
+export const Profile = () => {
+
+    return <div><h1>Профиль</h1></div>
+}
+
+export default Profile
