+++ src/routes/chats/chats.js
@@ -1,0 +1,16 @@
+import { Chat, ChatList } from '../../components'
+
+export const Chats = () => {
+    
+    return (
+    <>
+        <h1>
+            Чаты
+        </h1>
+
+        <Chat/>
+    </>
+    )
+}
+
+export default Chats
