+++ src/routes/home/home.js
@@ -1,0 +1,13 @@
+
+export const Home = () => {
+    
+    return (
+    <div>
+        <h1>
+            Home
+        </h1>
+    </div>
+    )
+}
+
+export default Home
