+++ src/routes/index.js
@@ -1,0 +1,3 @@
+export * from './chats/chats'
+export * from './home/home'
+export * from './profile/profile'
