+++ src/routes/profile/profile.js
@@ -1,0 +1,13 @@
+
+export const Profile = () => {
+    
+    return (
+    <div>
+        <h1>
+            Профиль
+        </h1>
+    </div>
+    )
+}
+
+export default Profile
